+++ src/GeolocationException.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Bkhim\Geolocation;
+
+/**
+ * Class GeolocationException.
+ *
+ * @author Adriano Rosa <https://adrianorosa.com>
+ * @date 2019-08-13 16:00
+ *
+ * @package Bkhim\Geolocation
+ */
+class GeolocationException extends \Exception
+{
+}