+++ src/GeolocationManager.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bkhim\GeoLocation;
+namespace Bkhim\Geolocation;
 
 use GeoIp2\Database\Reader;
 use GuzzleHttp\Client;
@@ -10,14 +10,14 @@ use InvalidArgumentException;
 use MaxMind\Db\Reader\InvalidDatabaseException;
 
 /**
- * Class GeoLocationManager.
+ * Class GeolocationManager.
  *
  * @author Adriano Rosa <https://adrianorosa.com>
  * @date 2019-08-13 11:50
  *
- * @package Bkhim\GeoLocation
+ * @package Bkhim\Geolocation
  */
-class GeoLocationManager
+class GeolocationManager
 {
     /**
      * @var array
@@ -35,7 +35,7 @@ class GeoLocationManager
     protected $cacheProvider;
 
     /**
-     * GeoLocation constructor.
+     * Geolocation constructor.
      *
      * @param  array  $config
      * @param  CacheManager  $cacheProvider
@@ -49,10 +49,10 @@ class GeoLocationManager
     }
 
     /**
-     * Get a GeoLocation driver instance.
+     * Get a Geolocation driver instance.
      *
      * @param string|null $name
-     * @return \Bkhim\GeoLocation\Contracts\LookupInterface
+     * @return \Bkhim\Geolocation\Contracts\LookupInterface
      */
     public function driver(?string $name = null): Contracts\LookupInterface
     {
@@ -90,20 +90,24 @@ class GeoLocationManager
      * Create IpInfo driver instance.
      *
      * @param  array $config
-     * @return \Bkhim\GeoLocation\Contracts\LookupInterface
+     * @return \Bkhim\Geolocation\Contracts\LookupInterface
      */
-    protected function createIpinfoDriver(array $config): Contracts\LookupInterface
+    protected function createIpinfoDriver($config)
     {
         $options = $config['client_options'] ?? [];
 
-        return new Providers\IpInfo(new Client($options), $this->cacheProvider->getStore());
+        // Use Laravel's HTTP client instead of raw Guzzle
+        return new Providers\IpInfo(
+            new \GuzzleHttp\Client($options),
+            $this->cacheProvider->getStore()
+        );
     }
 
     /**
      * Create MaxMind driver instance.
      *
      * @param  array  $config
-     * @return \Bkhim\GeoLocation\Contracts\LookupInterface
+     * @return \Bkhim\Geolocation\Contracts\LookupInterface
      * @throws InvalidArgumentException
      */
     protected function createMaxmindDriver(array $config): Contracts\LookupInterface
@@ -152,7 +156,7 @@ class GeoLocationManager
      * Get a provider instance.
      *
      * @param string|null $name
-     * @return \Bkhim\GeoLocation\Contracts\LookupInterface
+     * @return \Bkhim\Geolocation\Contracts\LookupInterface
      */
     protected function provider(?string $name = null): Contracts\LookupInterface
     {
@@ -164,7 +168,7 @@ class GeoLocationManager
      * Resolve the given provider.
      *
      * @param  string  $name
-     * @return \Bkhim\GeoLocation\Contracts\LookupInterface
+     * @return \Bkhim\Geolocation\Contracts\LookupInterface
      * @throws \InvalidArgumentException
      */
     protected function resolve(string $name): Contracts\LookupInterface
@@ -172,7 +176,7 @@ class GeoLocationManager
         $config = $this->getConfig($name);
 
         if (is_null($config)) {
-            throw new InvalidArgumentException("GeoLocation Driver [{$name}] is not defined.");
+            throw new InvalidArgumentException("Geolocation Driver [{$name}] is not defined.");
         }
 
         $driverMethod = 'create'.ucfirst($config['driver']).'Driver';
@@ -181,7 +185,7 @@ class GeoLocationManager
             return $this->{$driverMethod}($config);
         }
 
-        throw new InvalidArgumentException("GeoLocation Driver [{$config['driver']}] is not supported.");
+        throw new InvalidArgumentException("Geolocation Driver [{$config['driver']}] is not supported.");
     }
 
     /**